requirements.txt
line-bot-sdk
fastapi
uvicorn[standard]
langchain
openai
yfinance
pydantic
tiktoken
mysql